server/ocr/requirements.txt
paddlepaddle==2.6.1
paddleocr==2.8.1
PyMuPDF==1.24.1
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
regex==2024.4.28
beautifulsoup4==4.12.3
easyocr==1.7.1
lxml==5.2.1
python-dotenv==1.0.1
